# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type Comment struct {
	ID        int32         `json:"id"`
	PostID    int32         `json:"postId"`
	UserID    int32         `json:"userId"`
	Content   string        `json:"content"`
	ParentID  sql.NullInt32 `json:"parentId"`
	CreatedAt sql.NullTime  `json:"createdAt"`
	UpdatedAt sql.NullTime  `json:"updatedAt"`
}

type Follower struct {
	ID          int32        `json:"id"`
	FollowerID  int32        `json:"followerId"`
	FollowingID int32        `json:"followingId"`
	CreatedAt   sql.NullTime `json:"createdAt"`
}

type Like struct {
	ID        int32        `json:"id"`
	PostID    int32        `json:"postId"`
	UserID    int32        `json:"userId"`
	CreatedAt sql.NullTime `json:"createdAt"`
}

type Post struct {
	ID           int32         `json:"id"`
	UserID       int32         `json:"userId"`
	Content      string        `json:"content"`
	LikeCount    sql.NullInt32 `json:"likeCount"`
	ShareCount   sql.NullInt32 `json:"shareCount"`
	CommentCount sql.NullInt32 `json:"commentCount"`
	CreatedAt    sql.NullTime  `json:"createdAt"`
	UpdatedAt    sql.NullTime  `json:"updatedAt"`
}

type Share struct {
	ID        int32        `json:"id"`
	PostID    int32        `json:"postId"`
	UserID    int32        `json:"userId"`
	CreatedAt sql.NullTime `json:"createdAt"`
}

type User struct {
	ID             int32          `json:"id"`
	Username       string         `json:"username"`
	PasswordHash   string         `json:"passwordHash"`
	Bio            sql.NullString `json:"bio"`
	FollowersCount sql.NullInt32  `json:"followersCount"`
	FollowingCount sql.NullInt32  `json:"followingCount"`
	IsAdmin        sql.NullBool   `json:"isAdmin"`
	CreatedAt      sql.NullTime   `json:"createdAt"`
	UpdatedAt      sql.NullTime   `json:"updatedAt"`
	Origin         sql.NullString `json:"origin"`
}